pyproject.toml
[project]
name = "arelle-api"
version = "1.0.0"
description = "XBRL validation API using Arelle"
requires-python = ">=3.10"
dependencies = [
    "arelle-release>=2.37.74",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
]

[dependency-groups]
dev = []
